feat: Add .env configuration support and code cleanup
- Add .env and .env.example files with required configuration keys
- Implement BotConfig class for clean environment variable loading
- Add runtime assembly resolver for Newtonsoft.Json to fix startup crashes
- Rename PrvniStart to StartBot and improve log messages
- Clean up code structure and naming conventions
- Fix Newtonsoft.Json dependency issue without requiring MSBuild
