+++ .env.example
@@ -1,0 +1,11 @@
+﻿# EpicRPGBotCSharp environment configuration
+# Copy this file to `.env` and fill the values as needed.
+
+BASE_URL=https://discord.com
+DISCORD_CHANNEL_URL=http://discord.com/channels/0000000000000000000/0000000000000000000
+AREA=10
+DISCORD_EMAIL=
+DISCORD_PASSWORD=
+#Found in chrome://version
+SOURCE_USER_DATA_DIR="C:\Users\YOUR_USER\AppData\Local\Google\Chrome\User Data"
+PROFILE_NAME="Profile 1"
