--- .vs/test/v17/DocumentLayout.json
@@ -1,55 +1,0 @@
-{
-  "Version": 1,
-  "WorkspaceRootPath": "C:\\Projects\\test\\",
-  "Documents": [
-    {
-      "AbsoluteMoniker": "D:0:0:{D47BD117-13C9-4EC1-AD34-6626B86E5645}|test\\test.csproj|c:\\projects\\test\\test\\program.cs||{A6C744A8-0E4A-4FC6-886A-064283054674}",
-      "RelativeMoniker": "D:0:0:{D47BD117-13C9-4EC1-AD34-6626B86E5645}|test\\test.csproj|solutionrelative:test\\program.cs||{A6C744A8-0E4A-4FC6-886A-064283054674}"
-    },
-    {
-      "AbsoluteMoniker": "D:0:0:{00000000-0000-0000-0000-000000000000}|\u003CSolution\u003E|EpicRPGBotCSharp||{04B8AB82-A572-4FEF-95CE-5222444B6B64}|"
-    },
-    {
-      "AbsoluteMoniker": "D:0:0:{D47BD117-13C9-4EC1-AD34-6626B86E5645}|test\\test.csproj|c:\\projects\\test\\test\\||{B270807C-D8C6-49EB-8EBE-8E8D566637A1}|5e9a8ac2-4f34-4521-858f-4c248ba31532"
-    }
-  ],
-  "DocumentGroupContainers": [
-    {
-      "Orientation": 0,
-      "VerticalTabListWidth": 256,
-      "DocumentGroups": [
-        {
-          "DockedWidth": 200,
-          "SelectedChildIndex": 1,
-          "Children": [
-            {
-              "$type": "Document",
-              "DocumentIndex": 1,
-              "Title": "EpicRPGBotCSharp",
-              "DocumentMoniker": "C:\\Projects\\test\\test\\EpicRPGBotCSharp.csproj",
-              "RelativeDocumentMoniker": "test\\EpicRPGBotCSharp.csproj",
-              "ToolTip": "C:\\Projects\\test\\test\\EpicRPGBotCSharp.csproj",
-              "RelativeToolTip": "test\\EpicRPGBotCSharp.csproj",
-              "Icon": "ae27a6b0-e345-4288-96df-5eaf394ee369.001001|",
-              "WhenOpened": "2024-08-23T19:41:39.436Z",
-              "EditorCaption": ""
-            },
-            {
-              "$type": "Document",
-              "DocumentIndex": 0,
-              "Title": "Program.cs",
-              "DocumentMoniker": "C:\\Projects\\test\\test\\Program.cs",
-              "RelativeDocumentMoniker": "test\\Program.cs",
-              "ToolTip": "C:\\Projects\\test\\test\\Program.cs",
-              "RelativeToolTip": "test\\Program.cs",
-              "ViewState": "AgIAAEgAAAAAAAAAAAAAwFMAAAAJAAAAAAAAAA==",
-              "Icon": "ae27a6b0-e345-4288-96df-5eaf394ee369.000738|",
-              "WhenOpened": "2024-08-23T19:41:27.641Z",
-              "EditorCaption": ""
-            }
-          ]
-        }
-      ]
-    }
-  ]
-}
