--- test.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.10.35122.118
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EpicRPGBotCSharp", "test\EpicRPGBotCSharp.csproj", "{D47BD117-13C9-4EC1-AD34-6626B86E5645}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{D47BD117-13C9-4EC1-AD34-6626B86E5645}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D47BD117-13C9-4EC1-AD34-6626B86E5645}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D47BD117-13C9-4EC1-AD34-6626B86E5645}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D47BD117-13C9-4EC1-AD34-6626B86E5645}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {CD814384-95AA-409F-A07A-825B9A4E5DF3}
-	EndGlobalSection
-EndGlobal
